Rename profile photo app component
