+++ frontend/src/lib/components/App.svelte
