+++ frontend/src/main.js
@@ -1,7 +1,7 @@
 import "svelte-lib/styles/app.css"
 import "svelte-lib/styles/root.css"
 
-import ProfilePhoto from "./lib/components/ProfilePhoto.svelte"
+import ProfilePhoto from "./lib/components/App.svelte"
 
 let div = document.createElement("div")
 div.classList.add("profile-photo")
